Added a few methods to the PlaybackService-class. Refactored the MainActivity.
In the PlaybackService I created a showToast-method to display a message on the screen, overrided the onPositionDiscontinuity-method of the Player.Listener to determine whether the current item is last, overrided the onPositionDiscontinuity-method to determine whether the player is playing or on pause. In the MainActivity added a new media-item and refactored a logic which is responsible for adding tracks to the controller.